refactor : RequestHandler의 일부 메서드들을 utils와 HTTPResponseHandler로 옮겨 책임 분산
